backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
yfinance>=0.2.0
newsapi-python>=0.2.7
scikit-learn>=1.3.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
